refactor: move core to internal/core and adjust import aliases
