+++ adapters/primary/http/server/router.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"github.com/go-chi/chi/v5"
 	"github.com/lechitz/AionApi/adapters/primary/http/handlers"
-	inputHttp "github.com/lechitz/AionApi/core/ports/input/http"
-	outputHttp "github.com/lechitz/AionApi/core/ports/output/security"
+	inputHttp "github.com/lechitz/AionApi/internal/core/ports/input/http"
+	outputHttp "github.com/lechitz/AionApi/internal/core/ports/output/security"
 	"github.com/lechitz/AionApi/internal/platform/middleware/auth"
 	"go.uber.org/zap"
 )
