+++ cmd/aion-api/main.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	"github.com/lechitz/AionApi/adapters/primary/http/server"
-	"github.com/lechitz/AionApi/core/msg"
+	"github.com/lechitz/AionApi/internal/core/msg"
 	"github.com/lechitz/AionApi/internal/platform/bootstrap"
-	config2 "github.com/lechitz/AionApi/internal/platform/config"
+	"github.com/lechitz/AionApi/internal/platform/config"
 	"github.com/lechitz/AionApi/internal/platform/logger"
 	"github.com/lechitz/AionApi/pkg/contextkeys"
 	"github.com/lechitz/AionApi/pkg/errors"
@@ -17,13 +17,13 @@ func main() {
 
 	loggerSugar.Infow(msg.StartingApplication)
 
-	if err := config2.LoadConfig(loggerSugar); err != nil {
+	if err := config.LoadConfig(loggerSugar); err != nil {
 		errors.HandleCriticalError(loggerSugar, msg.ErrToFailedLoadConfiguration, err)
 		return
 	}
-	loggerSugar.Infow(msg.SuccessToLoadConfiguration, contextkeys.Setting, config2.Setting)
+	loggerSugar.Infow(msg.SuccessToLoadConfiguration, contextkeys.Setting, config.Setting)
 
-	appDependencies, cleanup, err := bootstrap.InitializeDependencies(loggerSugar, config2.Setting)
+	appDependencies, cleanup, err := bootstrap.InitializeDependencies(loggerSugar, config.Setting)
 	if err != nil {
 		errors.HandleCriticalError(loggerSugar, msg.ErrInitializeDependencies, err)
 		return
@@ -32,14 +32,14 @@ func main() {
 
 	loggerSugar.Infow(msg.SuccessToInitializeDependencies)
 
-	newServer, err := server.NewHTTPServer(appDependencies, loggerSugar, &config2.Setting)
+	newServer, err := server.NewHTTPServer(appDependencies, loggerSugar, &config.Setting)
 	if err != nil {
 		loggerSugar.Infow(msg.ErrStartServer)
 		errors.HandleCriticalError(loggerSugar, msg.ErrStartServer, nil)
 		return
 	}
 
-	loggerSugar.Infow(msg.ServerStarted, contextkeys.Port, newServer.Addr, contextkeys.ContextPath, config2.Setting.Server.Context)
+	loggerSugar.Infow(msg.ServerStarted, contextkeys.Port, newServer.Addr, contextkeys.ContextPath, config.Setting.Server.Context)
 
 	if err := newServer.ListenAndServe(); err != nil {
 		errors.HandleCriticalError(loggerSugar, msg.ErrStartServer, err)
