+++ internal/core/domain/token_domain.go
