+++ internal/core/msg/errors.go
