+++ internal/core/ports/output/db/user.go
@@ -1,7 +1,7 @@
 package db
 
 import (
-	"github.com/lechitz/AionApi/core/domain"
+	"github.com/lechitz/AionApi/internal/core/domain"
 )
 
 type IUserRepository interface {
