+++ internal/core/service/user_service.go
@@ -3,26 +3,26 @@ package service
 import (
 	"errors"
 	"fmt"
-	"github.com/lechitz/AionApi/core/ports/output/db"
-	security2 "github.com/lechitz/AionApi/core/ports/output/security"
+	"github.com/lechitz/AionApi/internal/core/domain"
+	"github.com/lechitz/AionApi/internal/core/msg"
+	"github.com/lechitz/AionApi/internal/core/ports/output/db"
+	"github.com/lechitz/AionApi/internal/core/ports/output/security"
 	"strings"
 	"time"
 
 	"github.com/badoux/checkmail"
-	"github.com/lechitz/AionApi/core/domain"
-	"github.com/lechitz/AionApi/core/msg"
 	"github.com/lechitz/AionApi/pkg/contextkeys"
 	"go.uber.org/zap"
 )
 
 type UserService struct {
 	UserRepository  db.IUserRepository
-	TokenService    security2.ITokenService
-	PasswordService security2.IPasswordService
+	TokenService    security.ITokenService
+	PasswordService security.IPasswordService
 	LoggerSugar     *zap.SugaredLogger
 }
 
-func NewUserService(userRepo db.IUserRepository, tokenService security2.ITokenService, passwordService security2.IPasswordService, loggerSugar *zap.SugaredLogger) *UserService {
+func NewUserService(userRepo db.IUserRepository, tokenService security.ITokenService, passwordService security.IPasswordService, loggerSugar *zap.SugaredLogger) *UserService {
 	return &UserService{
 		UserRepository:  userRepo,
 		TokenService:    tokenService,
