+++ internal/platform/bootstrap/bootstrap.go
@@ -4,13 +4,13 @@ import (
 	tokenadapter "github.com/lechitz/AionApi/adapters/secondary/cache"
 	dbadapter "github.com/lechitz/AionApi/adapters/secondary/db"
 	securityadapter "github.com/lechitz/AionApi/adapters/secondary/security"
-	"github.com/lechitz/AionApi/core/ports/input/http"
-	"github.com/lechitz/AionApi/core/ports/output/security"
+	"github.com/lechitz/AionApi/internal/core/ports/input/http"
+	"github.com/lechitz/AionApi/internal/core/ports/output/security"
+	"github.com/lechitz/AionApi/internal/core/service"
 	"github.com/lechitz/AionApi/internal/infrastructure/cache"
 	"github.com/lechitz/AionApi/internal/infrastructure/db"
 	"github.com/lechitz/AionApi/internal/platform/config"
 
-	"github.com/lechitz/AionApi/core/service"
 	"go.uber.org/zap"
 )
