+++ internal/platform/middleware/auth/auth_middleware.go
@@ -2,9 +2,9 @@ package auth
 
 import (
 	"context"
-	"github.com/lechitz/AionApi/core/domain"
-	inputHttp "github.com/lechitz/AionApi/core/ports/input/http"
-	outputHttp "github.com/lechitz/AionApi/core/ports/output/security"
+	"github.com/lechitz/AionApi/internal/core/domain"
+	inputHttp "github.com/lechitz/AionApi/internal/core/ports/input/http"
+	outputHttp "github.com/lechitz/AionApi/internal/core/ports/output/security"
 	"net/http"
 
 	"github.com/lechitz/AionApi/pkg/contextkeys"
