+++ tests/mocks/mock_token_service.go
@@ -5,10 +5,10 @@
 package mocks
 
 import (
-	reflect "reflect"
+	"github.com/lechitz/AionApi/internal/core/domain"
+	"reflect"
 
-	gomock "github.com/golang/mock/gomock"
-	domain "github.com/lechitz/AionApi/core/domain"
+	"github.com/golang/mock/gomock"
 )
 
 // MockITokenService is a mock of ITokenService interface.
